+++ src/client/graphical/static_animator.cpp
@@ -4,7 +4,7 @@ using namespace familyline::graphics;
 
 StaticAnimator::StaticAnimator(VertexDataGroup vdg) : _vdg(vdg) {}
 
-void StaticAnimator::advance(int ms) {}
+void StaticAnimator::advance(double ms) {}
 void StaticAnimator::runAnimation(const char* name) {}
 
 VertexDataGroup StaticAnimator::getCurrentFrame()
