+++ src/include/client/graphical/static_animator.hpp
@@ -28,12 +28,13 @@ public:
      * Advance animation by 'ms' milisseconds.
      * If ms is not the frametime of this animation, interpolate
      */
-    virtual void advance(int ms);
+    virtual void advance(double ms);
 
     /**
      * Run some predefined animation named 'name'.
      * Set the internal pointer to the first frame of that animation
      */
     virtual void runAnimation(const char* name);
+    
 };
 }  // namespace familyline::graphics
